fix(spi): changed uint8_t to uint16_t byte count in send/eceive functions

+++ drivers/SpiAbs/SpiAbs.h
@@ -23,7 +23,11 @@ uint8_t SpiAbs_CsDisable(enum SPIABS_DEVICE dev);
 
 uint8_t SpiAbs_readByte(enum SPIABS_DEVICE dev, uint8_t * resp);
 uint8_t SpiAbs_writByte(enum SPIABS_DEVICE dev, const uint8_t *data);
-uint8_t SpiAbs_SendReceiveMsg(enum SPIABS_DEVICE dev, const uint8_t * pTxBuffer, uint8_t * pRxBuffer, uint8_t TxBytes);
+uint8_t SpiAbs_SendReceiveMsg(
+    enum SPIABS_DEVICE dev, 
+    const uint8_t * pTxBuffer, 
+    uint8_t * pRxBuffer, 
+    uint16_t TxBytes);
 uint8_t SpiAbs_PollForResponse(enum SPIABS_DEVICE dev, uint8_t * pResponse);
 uint8_t SpiAbs_Send_Spi1_Task0(
     const uint8_t * data,
@@ -44,4 +48,9 @@ void __attribute__((weak)) SpiAbs_TaskControlCallback(uint32_t timeout);
 void __attribute__((weak)) SpiAbs_TaskSendReceiveCallback(void);
 
 
-void SpiAbs_Send_Spi1_CompleteCallback_Task0(void * context, uint32_t status, const uint8_t *data, uint32_t len);
+void SpiAbs_Send_Spi1_CompleteCallback_Task0(
+    void * context, 
+    uint32_t status, 
+    const uint8_t *data, 
+    uint32_t len
+);
